subsequences: init commit (adding tests for the upcoming program)
Adding tests for verification of the upcoming subsequences.py program.
A correct subsequences.py should pass all the tests.

How to run the tests against a tool:

./tests.sh your-tool-name.py

The file your-tool-name.py is expected to be executable and accept
two lines as input (stdin) and the longest subsequence should be
printed to stdout.