game15: init commit (adding tests for the upcoming program)
Adding tests for verification of the upcoming game15.py program.
A correct program should pass all the tests.

How to run the tests against your program:

./test.sh your-program.py

The your-program.py is expected to be executable and accept
a Game 15 position as input (stdin) and "YES" or "NO" is
expected to be printed back at the last line. It could be
the only line of the output, but the test script should
handle anything being printed before (steps, position, ...),
just the last line matters.

The testsuite contains a few resolvable and a few non-resolvable
positions stored in files YES_*.dat and NO_*.dat. It runs the
target program against all of the positions and expects the
results match the names of the files.