martingale.py: toggleable colors, toggleable result, added user-interaction
First I added toggleable colors everywhere. The color are easily toggled
with constant 'color'. They are also easily changed to fit a person's taste
by changing the color variables which always starts with a letter 'c'
(e.g. cTitle). The only colors that don't have a static color variable (but
are still toggleable) are 'cRoulette' color which are used when displaying
roulette game history in user-interaction section. A special color variable
would be 'cError' which is only used in user-interaction section when
printing error/warning message to the user. It has a predefined error
symbol, in this case '[!]'. The error symbol can be easily changed to other
symbol like for example to '[ERROR]' or '<oops>' or anything. That's one of
the reason why 'cError' is used only at the begining of the printed line.
The color variable are most self-explanitory but are not neccessary always
used only for one thing. For example 'cSubtitle' is also used for coloring
user's input at user-interaction section.

Next, I added toggleable in-depth result analyzation. I basically created a
'analysis' boolean variable and if statement (line 129).Having it False,
the program will exit after printing out basic result analyzation.

Finally the biggest change is the user-interaction section which is not
finished yet. The user-interaction starts at line 233 till the end of the
file. The main concept is to have the user to be able to search for specific
data that were not printed out. One of the purpose is to verify the
analyzation and accuracy of the program. The other purpose would be using
the "hidden" "useless" "realistic" elements (such as convert function) to a
good use. There are basically 3 core feature. First feature would be
printing out roulette history of specific rounds. Next would be printing
data on specific rounds and finally print data on specific players on
specific rounds. More information can be seen when typing 'help' in
user-interaction section. In order for some of the feature to work, the
program have to record ALL data from every single round. So if the
user-interaction is enabled (True) then the program will also collect
additional data (of every round from every player) into a multiple layered
list called 'data2'. The only commands are fully working right now in this
user-interaction section is help, history and quit/exit. Verifying arguments
for: the data searching based on player name or data searching based on round
number, is already implemented but the actual function is in the process of
development. In this commit the 'data2' recording is disabled (comment out)
because it is in a middle of development and it's in an error stage right
now (see comment '#ERROR').