subsequences.py: fix algorithm to pass 't6' test
Fix the program algorithm to be able to pass all tests including t6. The
program is still using a 'tmp' variable to store, add and remove characters
to compare to the other words. Before though, while the 'tmp' variable had
more then 1 character and when the condition that: 'tmp' is not in the other
words is met, it would reset the 'tmp' variable to blank and it would
continue with the next character that the program didn't try, meaning it
would skip the characters from 'tmp' (except from the first character). Now
it is fixed to go back and not skip any character.

EXAMPLE (using t6 test): the short word is 'cdehfge'. While looping, the
program find out that  'cde' is in the other word(s) but 'cdeh' is not in
other word(s) so the program would reset 'tmp' variable to blank and
continue with the character 'f' and skips 'd', 'e', and 'h'. Wih this fix,
the program will go back and won't skip.

The fix involves replacing the 'for' loop (that loops the short word
characters) to a 'while' loop. With that, we are able to freely move forward
and backwards in the loop freely using the index (variable 'i'). Next, when
the 'tmp' is not in the other word(s) condition is met, it would check
whether there is one or more character in the 'tmp' variable. If there is
only one character in 'tmp' then it would just simply reset the 'tmp'
variable to blank and continue like the previous program. But when there are
more then one characters in 'tmp', then it would deduct the index by the
number of character in 'tmp' and plus one (see line 18). By deducting the
index, we technically "go back in the loop". By adding 1, we don't go back
to the first character of 'tmp', since if we did, we would be stuck in an
endless loop.

If you want to see the working of the program, simply remove the '#DEV'
comments and when the program is working, it will print out what
character(s) it is testing and whether it is in the other word(s) or not.