game15: moving the stuff to is-solvable subdir
There will be two separate tasks handled within the game15 topic:
First, the current one (to read a game position and decide whether
the position can be solved using a sequence of legal moves or not)
and after that, some automated solver will be created in future.

I am moving the current testsuite into the is-solvable directory,
because the solver will have its own tests.