sudoku/c++: added a simple solving engine
This commit adds a simple solving engine to the C++ version. The
engine is able to solve the solve_tests/missing*.dat dumb puzzles
and also the solve_tests/easy_02.dat one. The easy_01 puzzle is
still too hard for it. "Too hard" means that there is no cell with
just one option. Current solving engine can solve only puzzles that
always have at least one empty cell with only one option to fill in.

To be able to solve puzzles with more than one option per cell, we'd
need to add some tree processing, which is to be added in future.