sudoku/c++: more intelligent engine
Added ability of branching in the tree, so now we don't give up when
we find a non-trivial position (a position where there are no cells
with only one option). Now, in a non-trivial position, the position
alongside with the _next_ move is stored into a move_stack and current
move is played. In case it fails (results into a position where a cell
with 0 options exists), the position is restored back from the stack,
including the _next_ move, which can be tried out...

Currently, the engine is able to solve also solving_tests/easy_01.dat
which is quite good. Harder puzzles are about to be added into the
tests soon.