sudoku/sudoku.py: Added function 'possibilities'
Function possibilities requires no argument and returns a 3D list
where the first layer is the board, the second layer are the rows
and the third layer are the tiles. the tiles can be None for tiles
that has already a number, and for tiles with zero aka tiles that
are blank/empty, there would be a list containing possible numbers
filled in that particular tile.

Also added function 'calculatePossible', which is essential for
function 'possibilities'. This argument is the particular tile
'row number' and 'column number'. With the argument, the function
will get the particular row, column and small box, calculate what
number are possible to fill and return the possible numbers in the
particular empty tile (list format).

Finally, also added function 'boxNumber' which converts the 'row
number' and 'column number' into 'small box number'. The function
will need to be rewrited due to the fact that it uses an 'ugly'
way to calculate the 'small box number'.