sudoku/sudoku.py: Fixing (added: optparse, constants, + small fixes)
Added OptionParser library for detecting optional arguments:
'--detail', '--color' and '--solve'.

Added constants for verify() function. Constants in the 100s are
'errors', constants in 200s are 'incorrect'.

Added a MSG dictionary to assosiate constant with a msg in a
string format.

Added function 'state' with one argument. Return the state (error,
incorrect, correct, solved) based on the constant number.

Modified dPrint function (remove try/except, now uses argument).

Removed function 'color'. Planned to add it back in the future.

Modified verify function to return constants instead of a string.

Modified what to print and what not to print in both mode (when
--detail argument is detected and not detected). Basically, first
it prints out the board the program received in a user-friendly
format, then it print the states. If the --solve argument is
detected then it prints out the solved board.

It also print 'solved' when it solves the puzzle so the test suite
'test_solving.sh' works.

Commented out printing board data in a list form.

Modified accordingly to work with constant number output from
verify function (some of the original lines are only commented
out and not removed).