sudoku/sudoku.py: Adding algorithm to solve hard sudoku (not working)
Added in-class variable 'backup', which stores backup data in a list
of tuples. Create the variable at class Game init.

In function 'genPossibilities()', set up to return False when
'calculatePossible()' returns empty list. It means that the sudoku
puzzle is wrong, therefore it should restore backup.

Created function 'morePossible()', get tileData from
'snallestPossibleNum()' function, and saves a backup of the
necessary datas.

Created function 'saveBackup()', accept argument tileData which is
a tuple of a empty tile coordinate (row number, column number) and
a list of possible number in the said tile. The function saves
necessary data into the in-class variable 'self.backup'. After
saving the data, it will then insert one of the possible number
into the empty tile.

Created function 'restoreBackup()', as its name imply, it restores
a previously made backup. It overwrites current board with a
backuped board. It then run 'saveBackup()' function.

Created function 'smallestPossibleNum()', it will output an empty
tile data (coordinates: row, column; list of possible numbers) with
the lowest numbers of possible number.

--------------------------------------------------
Solving 'easy' sudoku puzzle with 'onePossible()' function still
work. When solving the 'garder' sudoku puzzle, it does not work.
The program is prone to crash due to the fact that it tries to
restore a backup data eveb when there are no data backed up
anymore.