sudoku/sudoku.py: Fixing the algorithm that solves hard puzzles
When saving curent board into class variable 'self.backup', it
was saving as a reference and not as the data. Fixed it using
a function 'deepcopy()' from 'copy' library. It is not possible
to solve the issue using 'list()' or 'copy()'(<- from 'copy'
library) because the board is 2D.

Added an IF condition in function 'restoreBackup()', to check
if the class variable 'self.backup' is empty or not. If it is
empty, then it will be stated as 'impossible error', print the
board and exit.

Added more dPrint() for better debug:
	- dPrinting when tile does not have solution
	- dPrinting for impossible error
	- dPrinting when inserting number into tiles

Modifying few dPrint to print more information.
	- after saving or restoring backup, it also print a list
	  of possible numbers for the specific tile

Removed the cause of the issue #1:
>	if possibleNum == []:
>		return False

Removed the cause of the issue #2:
>	while not self.saveBackup(tileData):
>		pass

Fix the issue:
	In 'saveBackup()' function, check whether there are any
	more possible numbers in 'possibleNum' variable. In case
	there are no more, then it won't backup anything anymore.