number-validator/number-validator.py: Initial commit (experimental)
Brainstorming stage. Experimenting with various different stuff.

Created function 'gap()', returns the average number of the gap
between the numbers of a list. Accepts one argument which has to
be list.

Created function 'compare()' which compares targeted number with
a list. It compares using the 'howSimilar()' function and returns
the mean average. Accepts two arguments which are targeted number
and a list of numbers.

Created function 'howSimilar()'. Returns a float (0 - 1) which is
the similarity of the two number. Accepts two integers.