number-validator/number-validator.py: Upgrade (it's testable now)
Create function 'input()', no arguments, return the targeted
number and the list of numbers from the STDIN.

Function 'gap()' was renamed to 'gapInList()'. The function still
works the same, as it returns the average gap size between all
the numbers in the list.

Created function 'gapTargetNumber()' which is similar to the
function 'gapInList()'. It returns the average gap size between
the targeted number and the numbers in the list.

In 'howSimilar()' function, mathematically shortened the algorithm
to: abs(float(value_1)/float(value_2)).

Removed the debug testing lines.

Added a proper program output.