quadratic-solver/quadratic_eq_solver.c: Initial Commit (fully working)
Does not accept any sort of input whatsoever. Have to change the
variables 'a', 'b' and 'c' in the source code.

It then prints out the variable's definitions, full quadratic
equations with the variables in it, its discriminant and finally
the both (first and second) solutions.

It can also detect whether there are two solutions or only one
solution based on whether the discriminant equals zero or not.