brute-decrypt/brute_decryot.c: Initial commit (not completed yet)
First, it allocates STDIN into "encrMsg" with size 65536 through a
buffer. Then it enters a brute force loop which loops through 1-256.
In the brute-force loop is another loop looping through each character
of "encrMsg" and tries to decrypt it using the current key.

Missing function 'readability()' which should return a confidence
level (0-1) in how human-ish the msg is after decryption.