brute-decrypt/brute_decrypt.c: Repair & optimize (works for crypto)
Removed *CLEAR* makro since it is not necessary.

Function *readability()* returns *double* instead of *int* since the
*THRESHOLD* is a *double*. Thus variable *confidence* is *double*.

Function *readability()* accept another argument *size* as an *int*.

Commented out few instance where DEBUG msg were printed.

Removed entirely the character buffer (called *charBuf*) since it
is a) not necessary and b) stupid to use *strcat()* in the loop.

Added a variable (int) called *size* which stores the size of
STDIN. It is just to not have to run the *strlen()* function all
the time, which is unnecessary tax for the resources.

Added a check for the allocation of *encrMsg* variable. Note that
if that fail, it will free the *buf* buffer which was allocated
before.

At the end of the *main()* function, free everything (*buf*, *encrMsg*)
and *return 0;*.