brute-decrypt/brute_decrypt.c: Create function 'readability()'
Function 'readability()' now works and properly return the
'confidence' value as 'double'.

Include '<ctype.h>' to be able to perform lowercase function
'tolower()' in the 'readability()' function.

How function 'readability()' works:
    First it loops through characters of the buffer. It then
    lowercase each character. It check for each character ASCII
    value and see if the ASCII value is between the ASCII value of
    'a' and 'z'. If it is, then it would increment the 'counter'.
    Additionally, it also increment the 'counter' if it equals to
    space (' '). Finally it will calculate the confidence by
    divind the 'counter' with the 'size' of the whole string.

Added the feature to not print out debug messages by enclosing the
debug msg with the DIRECTIVE ('#ifdef (name of the makro)' and
'#endif'). To toggle on the debug printing: during compiling, use
'-D(name of the makro)'. In the script the name of the makro is
'DEBUG' --> so use '-DDEBUG' during compliation for debug msg.

Removed the header when printing decrypted msg.

Have all debug msg printed into STDERR. The key number is also
printed into STDERR though its not considered as debug msg.