brute-decrypt/brute_decrypt_v2.c: Initial commit (upgrading to v2)
Upgraded the original brute_decrypt.c from just using a predefined
THRESHOLD, into a script that checks all possible keys (1-255) and
determine the 'correct' key as the key with the most 'valid'
characters (which are letters from a-z and a space).

Added a STRUCTURE called 'best_t' which basically holds the best
key and number of character.

Instead of 'readability()' function, there is 'validChar()'
function which returns the number of 'valid' characers which
are letters from 'a' to 'z' and space.

Script works fine on most of the test sample however on the test
file crypto127 it truncated it.