brute-decrypt/brute_decrypt_v2.c: Fix truncation of some file
Change the way how to retrieve STDIN. Instead of using 'fgets();',
use 'fread();' which returns the size of what we read. This way we
don't need to use 'strlen();' to get the size of the STDIN, since
'strlen();' stops when there is a '\0' in the msg, which cause it
to truncate.