brute-decrypt/test.sh: Fix False-Negative
Removed entirely the variable $OUTPUT because when 'echoing' the
output, the output got process by 'echo' which can expand stuff,
merge double-spaces, adds newline...etc.

In short, the bash's echo process the output and thus modified
the output.

Fixed the issue by having the output be stored in a temporary
file by <(command)