sudoku/*: Add support to unsolvable sudoku (with no solution)
An "unsolvable" sudoku is referring to a sudoku that is correct
but does not have any solution.

Added an unsolvable sudoku to the test suite called
[unsolvable_01.dat] (which was created from hard_02).

Modified [test_solving.sh] to support unsolvable sudoku in the
test suite. If the file in the test suite folder (solving_tests)
starts with "unsolvable_", then it expect the program to output
"unsolvable". If any file in the test suite folder is named
differently, the test script assume it is solvable and has a
solution.

Modified [sudoku.py] to print appropriate message when there is
a sudoku that has no solution (aka unsolvable sudoku). Previously,
it printed "impossible error" and the current status of the board.
It was change so that when it stumble upon an unsolvable sudoku,
it prints the current status of the sudoku board first, following
with the message "unsolvable".